reverse-proxy: refactor/restructure codebase
